+++ LegendWeathers/Utils/Patches.cs
@@ -1,5 +1,6 @@
 ﻿using HarmonyLib;
 using LegendWeathers.BehaviourScripts;
+using LegendWeathers.Weathers;
 using System.Linq;
 
 namespace LegendWeathers.Utils
@@ -11,11 +12,11 @@ namespace LegendWeathers.Utils
         [HarmonyPatch("OnShipLandedMiscEvents")]
         public static void OnShipLandedMiscEventsPatch()
         {
-            // WILL CAUSE ISSUES UNTIL VANILLA IS FIXED
-            /*if (Effects.IsWeatherEffectPresent("majoramoon"))
+            string title = "Weather alert!";
+            if (Effects.IsWeatherEffectPresent("majoramoon"))
             {
-                Effects.MessageOneTime("Weather alert!", MajoraMoonWeather.weatherAlert, true, "LW_MajoraTip");
-            }*/
+                Effects.MessageOneTime(title, MajoraMoonWeather.weatherAlert, true, "LW_MajoraTip");
+            }
         }
     }